added search for sales

+++ src/main/java/ru/softwerke/controller/search/ClientSearchController.java
@@ -14,7 +14,7 @@ import java.util.stream.Stream;
 public class ClientSearchController extends InitController {
 
     public void searchById(String id) {
-        long searchId = Long.parseLong(id);
+        int searchId = Integer.parseInt(id);
         ClientList.getClientStream()
                 .filter(s -> (s.getClientId() == searchId))
                 .forEach(c -> ReadWriter.printLine(c));
